(zad 2) fully completed (enable operating on sales offer from the console)

+++ src/main/java/OfertaSprzedazy.java
@@ -9,7 +9,7 @@ import lombok.*;
 public class OfertaSprzedazy {
 
     private String nazwaProduktu;
-    private int cena;
+    private Double cena;
 
     @Override
     public String toString() {
